requirements.txt
flask
firebase_admin
line-bot-sdk
python-dotenv
gunicorn
